myWallet/requirements.txt
djangorestframework
json
traceback
sys
json
wsgi